move to app internal-package, rename rootFields -> topLeveLfields
