+++ src/internal-packages/app/index.js
@@ -3,6 +3,7 @@ const app = require('hadron-app');
 const InstanceActions = require('./lib/actions/instance-actions');
 const InstanceStore = require('./lib/stores/instance-store');
 const CollectionStore = require('./lib/stores/collection-store');
+const FieldStore = require('./lib/stores/field-store');
 const { NamespaceStore } = require('hadron-reflux-store');
 
 /**
@@ -13,6 +14,7 @@ function activate() {
   app.appRegistry.registerStore('App.InstanceStore', InstanceStore);
   app.appRegistry.registerStore('App.CollectionStore', CollectionStore);
   app.appRegistry.registerStore('App.NamespaceStore', NamespaceStore);
+  app.appRegistry.registerStore('Schema.FieldStore', FieldStore);
 }
 
 /**
@@ -23,6 +25,7 @@ function deactivate() {
   app.appRegistry.deregisterStore('App.InstanceStore');
   app.appRegistry.deregisterStore('App.CollectionStore');
   app.appRegistry.deregisterStore('App.NamespaceStore');
+  app.appRegistry.deregisterStore('Schema.FieldStore');
 }
 
 module.exports.activate = activate;
