--- src/internal-packages/schema-light/README.md
@@ -1,11 +1,0 @@
-# Compass Schema Package
-
-Executes a "light" schema analysis on documents fetched by the CRUD document stores.
-
-## Available Resources in the App Registry
-
-### Stores
-
-| Key                 | Description                                             |
-|---------------------|---------------------------------------------------------|
-| `Schema.FieldStore` | Triggers with sampling information about schema fields. |
